# [SUGEST] - A sugest for reduce further the script