--- a/src/index.js
+++ b/src/index.js
@@ -1,26 +1,25 @@
-function toChars(str) {
-	var i=0, arr=[];
-	while (i < str.length) {
-		arr.push(str.charCodeAt(i++));
-	}
-	return arr;
+function toChars (str) {
+	return str.split('').map((n, i) => str.charCodeAt(i))
 }
 
-function reduce(hasher, base, value) {
-	var i=0, j, tmp, str='';
-	var chars = base ? value.match(/.{1,2}/g) : toChars(value);
+function reduce (hasher, base, value) {
+	let i = 0
+	let str = ''
+	const chars = base ? value.match(/.{1,2}/g) : toChars(value)
+
 	while (i < chars.length) {
-		tmp = base ? parseInt(chars[i++], 16) : chars[i++];
-		for (j=0; j < hasher.length;) tmp ^= hasher[j++];
-		str += base ? String.fromCharCode(tmp) : ('0' + tmp.toString(16)).substr(-2);
+		let tmp = base ? parseInt(chars[i++], 16) : chars[i++]
+		for (let j = 0; j < hasher.length;) tmp ^= hasher[j++]
+		str += base ? String.fromCharCode(tmp) : ('0' + tmp.toString(16)).substr(-2)
 	}
-	return str;
+
+	return str
 }
 
-export function encrypt(key) {
-	return reduce.bind(reduce, toChars(key), 0);
+export function encrypt (key) {
+	return reduce.bind(reduce, toChars(key), 0)
 }
 
-export function decrypt(key) {
-	return reduce.bind(reduce, toChars(key), 1);
+export function decrypt (key) {
+	return reduce.bind(reduce, toChars(key), 1)
 }