Implemented `map()` to work for convert string to char. Has a revert, I was implement es6 too but I see that es5 is proprosital.