@luan0ap 2019-03-30
> Hey, thanks~!
> 
> I had this initially, but the manual loop is much faster as it only traversing the string once. I felt that the significant performance increase was worth the additional three bytes. What do you think?
> 
> Here's a before/after comparison:
> 
> ```js
> // Before:   259B gzip
> encrypt   x 288,653 ops/sec ±0.38% (93 runs sampled)
> decrypt   x 405,658 ops/sec ±1.73% (93 runs sampled)
> 
> // After:  256B gzip
> encrypt   x 241,159 ops/sec ±0.38% (94 runs sampled)
> decrypt   x 353,091 ops/sec ±1.48% (97 runs sampled)
> ```

Indeed, always will be `while` and `for` structure more fast, but i think, that the code is more readable using `map` or/and `forEach`. In this case, if you wanna brush bits, is sure use `while` for this.